refactor significantly the cli and split up the large file
