+++ README.md
@@ -57,7 +57,7 @@ The OceanDB package provides a CLI for initializing the database and ingesting d
 
 2. **Downloading Along-Track Data**
 
-Use `oceandb download` to fetch Copernicus Marine Service along-track data.
+Use `oceandb download along-track` to fetch Copernicus Marine Service along-track data.
 The command downloads files into `ALONG_TRACK_DATA_DIRECTORY`, which must be
 set in `.env`.
 
@@ -74,32 +74,32 @@ before downloading anything.
 Preview a download without fetching files:
 
 ```bash
-oceandb download --dry-run j3 --start-date 2024-01-01 --end-date 2024-02-01
+oceandb download along-track --dry-run j3 --start-date 2024-01-01 --end-date 2024-02-01
 ```
 
 Download one mission. The command previews the matching files, then prompts
 for confirmation:
 
 ```bash
-oceandb download j3 --start-date 2024-01-01 --end-date 2024-02-01
+oceandb download along-track j3 --start-date 2024-01-01 --end-date 2024-02-01
 ```
 
 Download multiple missions:
 
 ```bash
-oceandb download s3a s3b --start-date 2024-01-01 --end-date 2024-02-01
+oceandb download along-track s3a s3b --start-date 2024-01-01 --end-date 2024-02-01
 ```
 
 Download all supported along-track missions:
 
 ```bash
-oceandb download all --dataset-version 202411
+oceandb download along-track all --dataset-version 202411
 ```
 
 Skip the confirmation prompt for scripted runs:
 
 ```bash
-oceandb download j3 --start-date 2024-01-01 --end-date 2024-02-01 --yes
+oceandb download along-track j3 --start-date 2024-01-01 --end-date 2024-02-01 --yes
 ```
 
 After data has been downloaded, ingest it into OceanDB.
@@ -115,7 +115,7 @@ oceandb visualize-basins --output artifacts/basin_map.html
 
 3. **Downloading Eddy Data**
 
-Use `oceandb download-eddy` to fetch the packaged AVISO eddy NetCDF files.
+Use `oceandb download eddy` to fetch the packaged AVISO eddy NetCDF files.
 The command downloads files into `EDDY_DATA_DIRECTORY` and requires AVISO
 credentials in `.env`.
 
@@ -132,31 +132,31 @@ already exist locally. By default, it skips existing files unless
 Preview the download without fetching files:
 
 ```bash
-oceandb download-eddy --dry-run
+oceandb download eddy --dry-run
 ```
 
 Download the missing eddy files:
 
 ```bash
-oceandb download-eddy
+oceandb download eddy
 ```
 
 Skip the confirmation prompt for scripted runs:
 
 ```bash
-oceandb download-eddy --yes
+oceandb download eddy --yes
 ```
 
 Force a re-download of files that already exist locally:
 
 ```bash
-oceandb download-eddy --overwrite --yes
+oceandb download eddy --overwrite --yes
 ```
 
 4. **Ingesting Along-Track Data**
 
-`oceandb ingest-along-track` reads from the same
-`ALONG_TRACK_DATA_DIRECTORY` used by `oceandb download`.
+`oceandb ingest along-track` reads from the same
+`ALONG_TRACK_DATA_DIRECTORY` used by `oceandb download along-track`.
 Set `OCEANDB_INGEST_MODE=copy` in `.env` to stage batches through PostgreSQL
 `COPY`; the default remains `insert`.
 
@@ -164,19 +164,19 @@ By default if no arguments are provided this CLI command will iterate over all
 of the data.
 
 ```bash
-oceandb ingest-along-track
-oceandb ingest-along-track s3a
-oceandb ingest-along-track s3a j3 c2
-oceandb ingest-along-track j3 --start-date 2019-01-01 --end-date 2020-12-03
-oceandb ingest-along-track s6a --end-date 2024-01-01
-oceandb ingest-along-track s6a --start-date 2024-01-01
-oceandb summary alongtrack
+oceandb ingest along-track
+oceandb ingest along-track s3a
+oceandb ingest along-track s3a j3 c2
+oceandb ingest along-track j3 --start-date 2019-01-01 --end-date 2020-12-03
+oceandb ingest along-track s6a --end-date 2024-01-01
+oceandb ingest along-track s6a --start-date 2024-01-01
+oceandb summary along-track
 ```
 
 
 Ingesting Eddy Data
 ```bash
-oceandb ingest-eddy
+oceandb ingest eddy
 ```
